Move prepare_lembrame to process.py to be able to rewrite the package list
